# utils/documentation/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Primera versión de la API para la plataforma tecnológica de SAHEYA. Próxima de lanzamiento.",
    "version": "1.0.0",
    "title": "SAHEYA GUANLI",
    "termsOfService": "none",
    "contact": {
      "email": "feijizoule@126.com"
    }
  },
  "basePath": "/api",
...
  "paths": {
...
    "/loan/{loanId}": {
      "get": {
        "tags": [
          "loans"
        ],
        "summary": "Get one loan data",
        "description": "Admin users can get all loans data, but regular users can get only their loans under their same user id.",
        "operationId": "getLoan",
        "parameters": [
          {
            "name": "loanId",
            "in": "path",
            "description": "ID of the loan to return",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Loan",
            "schema": {
              "$ref": "#/definitions/Loan"
            }
          }
        }
      },
      "put": {
        "tags": [
          "loans"
        ],
        "summary": "Aprove - reject co-signed loans.",
        "description": "Use this method to approve or reject co-signed loans. Users can only modify loans under the respective account.",
        "operationId": "approveLoanCoSing",
        "parameters": [
          {
            "name": "loanId",
            "in": "path",
            "description": "ID loan to be modified",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "body",
            "in": "body",
            "description": "contains data to make the operation",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Loan"
            }
          }
        ]
      }
    }
  },
  "definitions": {
...
    "Loan": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "fecha_inicial": {
          "type": "integer",
          "format": "int64"
        },
        "num_cuotas": {
          "type": "integer",
          "format": "int64"
        },
        "deudor_id": {
          "type": "integer",
          "format": "int64"
        },
        "estado": {
          "type": "integer",
          "format": "int64"
        },
        "monto": {
          "type": "integer",
          "format": "int64"
        },
        "fecha_registro": {
          "type": "integer",
          "format": "int64"
        },
        "pagado": {
          "type": "integer",
          "format": "int64"
        }
      }
    }
  }
}